+++ onlinejudge/mathematics/adhoc/P996.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <algorithm>
 
-typedef int32_t ll;
+typedef int64_t ll;
 using namespace std;
 
 int main() {
